+++ src/main/java/net/silentchaos512/scalinghealth/datagen/GenEvent.java
@@ -12,7 +12,8 @@ public class GenEvent {
     @SubscribeEvent
     public static void gatherData(GatherDataEvent event){
         DataGenerator generator = event.getGenerator();
-        generator.addProvider(new Recipes(generator));
-        generator.addProvider(new LootTables(generator));
+        generator.addProvider(new SHTags(generator));
+        //generator.addProvider(new Recipes(generator));
+        //generator.addProvider(new LootTables(generator));
     }
 }