+++ src/main/java/net/silentchaos512/scalinghealth/init/ModLoot.java
@@ -25,6 +25,7 @@ import net.minecraft.world.storage.loot.conditions.LootConditionManager;
 import net.minecraft.world.storage.loot.functions.SetCount;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.LootTableLoadEvent;
+import net.silentchaos512.scalinghealth.loot.conditions.EntityGroupCondition;
 import net.silentchaos512.scalinghealth.loot.conditions.SHMobProperties;
 import net.silentchaos512.scalinghealth.utils.EnabledFeatures;
 
@@ -36,6 +37,7 @@ public final class ModLoot {
     public static void init() {
         MinecraftForge.EVENT_BUS.addListener(ModLoot::onLootTableLoad);
 
+        LootConditionManager.registerCondition(EntityGroupCondition.SERIALIZER);
         LootConditionManager.registerCondition(SHMobProperties.SERIALIZER);
     }