+++ src/main/java/net/silentchaos512/scalinghealth/lib/EntityGroup.java
@@ -14,18 +14,19 @@ import javax.annotation.Nullable;
 import java.util.Optional;
 
 public enum EntityGroup {
-    PEACEFUL("peaceful"),
-    HOSTILE("hostile"),
-    BOSS("boss"),
-    BLIGHT(null),
-    PLAYER(null);
+    PEACEFUL,
+    HOSTILE,
+    BOSS,
+    BLIGHT,
+    PLAYER;
 
-    private final ResourceLocation bonusDropsLootTable;
-
-    EntityGroup(@Nullable String bonusDropsTable) {
-        bonusDropsLootTable = bonusDropsTable != null
-                ? new ResourceLocation(ScalingHealth.MOD_ID, "bonus_drops/" + bonusDropsTable)
-                : null;
+    public static EntityGroup from(String name) {
+        for(EntityGroup group : values()){
+            if(group.name().equalsIgnoreCase(name)){
+                return group;
+            }
+        }
+        throw new RuntimeException("Could not get an entity group from name: " + name);
     }
 
     public static EntityGroup from(LivingEntity entity) {
@@ -53,10 +54,6 @@ public enum EntityGroup {
 
     public MobHealthMode getHealthMode() {
         return SHMobs.healthMode();
-    }
-
-    public Optional<ResourceLocation> getBonusDropsLootTable() {
-        return Optional.ofNullable(bonusDropsLootTable);
     }
 
     public boolean isAffectedByDamageScaling() {