+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+##[2.5.6] - 2020-05-15
+### Fixed
+- Backported new tag system for blight/difficulty exemption.
+- Mobs now use global loot modifier system. 
+
 ##[2.5.5] - 2020-05-15
 First release of 1.14.4 version! Yay
 ### Added